+++ src/full/Agda/TypeChecking/Monad/Options.hs
@@ -11,7 +11,6 @@ import Control.Monad.State  hiding (mapM)
 import Data.Maybe
 import Data.Traversable
 
-import Text.PrettyPrint
 import System.Directory
 import System.FilePath
 
@@ -34,6 +33,7 @@ import Agda.Utils.Maybe
 import Agda.Utils.Monad
 import Agda.Utils.Lens
 import Agda.Utils.List
+import Agda.Utils.Pretty
 import Agda.Utils.Trie (Trie)
 import qualified Agda.Utils.Trie as Trie
 import Agda.Utils.Except
@@ -196,12 +196,26 @@ setIncludeDirs incs relativeTo = do
       -- Might also be useful to overwrite default imports...
   incs <- return $ incs ++ [primdir]
 
+  reportSDoc "setIncludeDirs" 10 $ return $ vcat
+    [ text "Old include directories:"
+    , nest 2 $ vcat $ map pretty oldIncs
+    , text "New include directories:"
+    , nest 2 $ vcat $ map pretty incs
+    ]
+
   -- Check whether the include dirs have changed.  If yes, reset state.
   -- Andreas, 2013-10-30 comments:
   -- The logic, namely using the include-dirs variable as a driver
   -- for the interaction, qualifies for a code-obfuscation contest.
   -- I guess one Boolean more in the state cost 10.000 EUR at the time
   -- of this implementation...
+  --
+  -- Andreas, perhaps you have misunderstood something: If the include
+  -- directories have changed and we do not reset the decoded modules,
+  -- then we might (depending on how the rest of the code works) end
+  -- up in a situation in which we use the contents of the file
+  -- "old-path/M.agda", when the user actually meant
+  -- "new-path/M.agda".
   when (oldIncs /= incs) $ do
     ho <- getInteractionOutputCallback
     resetAllState